msg: add CEPH_MSG_FOOTER_NOHEADERCRC footer flag.
It is set when crc for footer headers is calculated. Check crc for
footer headers only if this flag is set.

Note, this introduces some undesirable effect if older mons are in the
cluster: as they do not set NOHEADERCRC flag, crc for footer headers
from these mons will not be checked by new mons.

Also rename CEPH_MSG_FOOTER_NOCRC to CEPH_MSG_FOOTER_NODATACRC, which
is more correct.

Signed-off-by: Mykola Golub <mgolub@mirantis.com>