msg: use flag in header to indicate the compression.
A flag in message header can indicate which parts of messages (payload, middle
or data) need to be compressed. The first 4 bytes of each of these payloads are
the original uncompressed length, and the next 4 bytes are the original CRC.
This implementation minimize the need to change fundamental msg structures.

Signed-off-by: Cheng Cheng <ccheng.leo@gmail.com>